requirements.txt
django
djangorestframework
faker
django-rest-knox